refactor email
