+++ pipelines/check_fcast_trigger.py
@@ -1,8 +1,8 @@
-from src.email import email, plotting
+from src.email import plotting, update_emails
 from src.monitoring import monitoring_utils
 
 if __name__ == "__main__":
     monitoring_utils.update_fcast_monitoring()
-    email.update_fcast_trigger_emails()
+    update_emails.update_fcast_trigger_emails()
     plotting.update_fcast_plots()
-    email.update_fcast_info_emails()
+    update_emails.update_fcast_info_emails()
