+++ src/email/send_emails.py
@@ -6,7 +6,6 @@ from email.message import EmailMessage
 from email.utils import make_msgid
 from typing import Literal
 
-import pandas as pd
 import pytz
 from html2text import html2text
 from jinja2 import Environment, FileSystemLoader
@@ -21,198 +20,13 @@ from src.email.utils import (
     EMAIL_USERNAME,
     STATIC_DIR,
     TEMPLATES_DIR,
-    TEST_ATCF_ID,
-    TEST_LIST,
     TEST_MONITOR_ID,
     TEST_STORM,
     add_test_row_to_monitoring,
+    get_distribution_list,
 )
 from src.monitoring import monitoring_utils
 from src.utils import blob
-
-
-def get_distribution_list() -> pd.DataFrame:
-    """Load distribution list from blob storage."""
-    if TEST_LIST:
-        blob_name = f"{blob.PROJECT_PREFIX}/email/test_distribution_list.csv"
-    else:
-        blob_name = f"{blob.PROJECT_PREFIX}/email/distribution_list.csv"
-    return blob.load_csv_from_blob(blob_name)
-
-
-def load_email_record() -> pd.DataFrame:
-    """Load record of emails that have been sent."""
-    blob_name = f"{blob.PROJECT_PREFIX}/email/email_record.csv"
-    return blob.load_csv_from_blob(blob_name)
-
-
-def update_fcast_info_emails(verbose: bool = False):
-    df_monitoring = monitoring_utils.load_existing_monitoring_points("fcast")
-    df_existing_email_record = load_email_record()
-    if TEST_STORM:
-        df_monitoring = add_test_row_to_monitoring(df_monitoring, "fcast")
-        df_existing_email_record = df_existing_email_record[
-            ~(
-                (df_existing_email_record["atcf_id"] == TEST_ATCF_ID)
-                & (df_existing_email_record["email_type"] == "info")
-            )
-        ]
-
-    dicts = []
-    for monitor_id, row in df_monitoring.set_index("monitor_id").iterrows():
-        if (
-            monitor_id
-            in df_existing_email_record[
-                df_existing_email_record["email_type"] == "info"
-            ]["monitor_id"].unique()
-        ):
-            if verbose:
-                print(f"already sent info email for {monitor_id}")
-        else:
-            try:
-                print(f"sending info email for {monitor_id}")
-                send_info_email(monitor_id=monitor_id, fcast_obsv="fcast")
-                dicts.append(
-                    {
-                        "monitor_id": monitor_id,
-                        "atcf_id": row["atcf_id"],
-                        "email_type": "info",
-                    }
-                )
-            except Exception as e:
-                print(f"could not send info email for {monitor_id}: {e}")
-
-    df_new_email_record = pd.DataFrame(dicts)
-    df_combined_email_record = pd.concat(
-        [df_existing_email_record, df_new_email_record], ignore_index=True
-    )
-    blob_name = f"{blob.PROJECT_PREFIX}/email/email_record.csv"
-    blob.upload_csv_to_blob(blob_name, df_combined_email_record)
-
-
-def update_obsv_trigger_emails():
-    df_monitoring = monitoring_utils.load_existing_monitoring_points("obsv")
-    df_existing_email_record = load_email_record()
-    if TEST_STORM:
-        df_monitoring = add_test_row_to_monitoring(df_monitoring, "obsv")
-        df_existing_email_record = df_existing_email_record[
-            ~(
-                (df_existing_email_record["atcf_id"] == TEST_ATCF_ID)
-                & (df_existing_email_record["email_type"] == "obsv")
-            )
-        ]
-    dicts = []
-    for atcf_id, group in df_monitoring.groupby("atcf_id"):
-        if (
-            atcf_id
-            in df_existing_email_record[
-                df_existing_email_record["email_type"] == "obsv"
-            ]["atcf_id"].unique()
-        ):
-            print(f"already sent obsv email for {atcf_id}")
-        elif (
-            atcf_id
-            in df_existing_email_record[
-                df_existing_email_record["email_type"] == "action"
-            ]["atcf_id"].unique()
-            and not TEST_STORM
-        ):
-            print(f"already sent action email for {atcf_id}")
-        else:
-            for monitor_id, row in group.set_index("monitor_id").iterrows():
-                if row["obsv_trigger"]:
-                    try:
-                        print(f"sending obsv email for {monitor_id}")
-                        send_trigger_email(
-                            monitor_id=monitor_id, trigger_name="obsv"
-                        )
-                        dicts.append(
-                            {
-                                "monitor_id": monitor_id,
-                                "atcf_id": atcf_id,
-                                "email_type": "obsv",
-                            }
-                        )
-                    except Exception as e:
-                        print(
-                            f"could not send trigger email for {monitor_id}: "
-                            f"{e}"
-                        )
-
-    df_new_email_record = pd.DataFrame(dicts)
-    df_combined_email_record = pd.concat(
-        [df_existing_email_record, df_new_email_record], ignore_index=True
-    )
-    blob_name = f"{blob.PROJECT_PREFIX}/email/email_record.csv"
-    blob.upload_csv_to_blob(blob_name, df_combined_email_record)
-
-
-def update_fcast_trigger_emails():
-    """Cycle through all historical monitoring points to see if we should have
-    sent a trigger email for any of them. If we need to send any emails,
-    send them.
-    """
-    df_monitoring = monitoring_utils.load_existing_monitoring_points("fcast")
-    df_existing_email_record = load_email_record()
-    if TEST_STORM:
-        df_monitoring = add_test_row_to_monitoring(df_monitoring, "fcast")
-        df_existing_email_record = df_existing_email_record[
-            ~(
-                (df_existing_email_record["atcf_id"] == "TEST_ATCF_ID")
-                & (
-                    df_existing_email_record["email_type"].isin(
-                        ["readiness", "action"]
-                    )
-                )
-            )
-        ]
-    dicts = []
-    for atcf_id, group in df_monitoring.groupby("atcf_id"):
-        for trigger_name in ["readiness", "action"]:
-            if (
-                atcf_id
-                in df_existing_email_record[
-                    df_existing_email_record["email_type"] == trigger_name
-                ]["atcf_id"].unique()
-            ):
-                print(f"already sent {trigger_name} email for {atcf_id}")
-            else:
-                for (
-                    monitor_id,
-                    row,
-                ) in group.set_index("monitor_id").iterrows():
-                    if (
-                        row[f"{trigger_name}_trigger"]
-                        and not row["past_cutoff"]
-                    ):
-                        try:
-                            print(
-                                f"sending {trigger_name} email for "
-                                f"{monitor_id}"
-                            )
-                            send_trigger_email(
-                                monitor_id=monitor_id,
-                                trigger_name=trigger_name,
-                            )
-                            dicts.append(
-                                {
-                                    "monitor_id": monitor_id,
-                                    "atcf_id": atcf_id,
-                                    "email_type": trigger_name,
-                                }
-                            )
-                        except Exception as e:
-                            print(
-                                f"could not send trigger email for "
-                                f"{monitor_id}: {e}"
-                            )
-
-    df_new_email_record = pd.DataFrame(dicts)
-    df_combined_email_record = pd.concat(
-        [df_existing_email_record, df_new_email_record], ignore_index=True
-    )
-    blob_name = f"{blob.PROJECT_PREFIX}/email/email_record.csv"
-    blob.upload_csv_to_blob(blob_name, df_combined_email_record)
 
 
 def send_info_email(monitor_id: str, fcast_obsv: Literal["fcast", "obsv"]):
