+++ src/email/templates/informational.html
@@ -16,7 +16,7 @@
     Déclencheur d'action: <b>{{ action }}</b>
   </p>
   <p>
-    Un tracé des {{ fcast_obsv }} est présenté ci-dessous.
+    Un tracé des {{ fcast_obsv }}s est présenté ci-dessous.
   </p>
   <img src="cid:{{ map_cid }}" width="600"/>
